Deck generator.  Related to #4
